Improve docking waypoint display
Fixes a display issue whereby the active docking waypoint vecdraw was calculated incorrectly (didn't take into account the displacement between the centre of mass of the ship and the docking port's face). Also changed some of the text and colours. Should be no difference in functionality (ultimately, the vecdraws and print outs could be removed, as was done for rendezvous).