Update lib_dock.ks
Replace calls to `enableRCS()` and `disableRCS()` with `toggleRCS()` (issue #69).
Add third waypoint if necessary (issue #18)
Tweak to initial placement of close node (may fix issue #17)
Rearrange code so that draw ing waypoints is in a separate function. Colour obstructed steps red.
Abort docking if the route becomes obstructed or if we run low on monoprop.