Update lib_orbit_match.ks
Trying three different methods of calculating. They should all return the same value, in which case we can use the shortest one and update the documentation for this library.

I think I now know what the vector manipulation in nodeMatchAtNode does to calculate the delta-v required for an inclination change. It boils down to calculating the length of the third side of an equilateral triangle whose other sides have a length proportional to the ship's velocity. As such, I am still at a loss as to why my first attempt using the law of cosines didn't work. It should return the same...