Update lib_transfer.ks
Node scoring - make bonus points for being close to target periapsis smaller, and follow a range (3 for perfection, dropping to 1 at 2500m difference, dropping to 0 at 50000m difference) rather than a fixed value.

`doTransfer()` - break out the check for whether we need a correction or not into another function. Change the periapsis accuracy we'll allow depending on how low the target periapsis is. If the target periapsis is within 15km of the atmosphere or below 20km, the allowed periapsis difference is 1km if in the current SoI, 10km if in the next SoI, 100km in next etc. If the target periapsis is below 250km, the allowed differences are 10km, 100km, 1000km etc. Above 250km, the allowed differences are 25km, 250km, 2500km etc.