Create lib_rcs_burn.ks
First draft of a lilbrary for burning a manoeuvre node with just RCS thrusters (though neither this nor `lib_burn` would currently handle the OMS engine that burns monoprop).

The big question that needs resolving is how to determine the ISP and thrust of the available thrusters. A guess is currently made that the ISP is 250s and that there is 1kN of thrust available.