Another calculation update
A better calculation for the distance required to decelerate, taking into account non-constant acceleration.
This is only used for the initial calculation of the burn time, for now. As it requires the lib_dv functions to calculate burn time, it may be too much processing to be used each frame during descent itself.