Don't always call `setIspFuelRate()`
Calling `setIspFuelRate()` is now optional when calling `burnTime()` and `stageDV()`.
We also now set the global variable `LIB_DV_LOADED` (initialised to false by `init_common.ks`) to true. This is so that the `setIspFuelRate()` function can be called by `doStage()`, but only if `lib_dv.ks` has been loaded.