Update lib_orbit.ks
Add `radiusAt()` and logging to `posAt()` to compare the size of the two. I think that `posAt()` only works if we are requesting a position vector at a time we are orbiting the same body, or the parent of that body. It's definitely the case that it goes weird when calculating a position from Kerbin while in orbit of the Mun. Ideally we would fix `posAt()`, but for now this should provide a good workaround.