Fix small burns
The changes for #121 broke the small burn calculation - this change
should fix it.

Basically: we need to recalculate the fuel rate, thrust etc. if we
decide to use a different throttle limiter compared to the previous
calculation.

That said, I'm still seeing small burns fail to burn a sizeable
proportion of the node's delta-v. I'm still investigating that....