Handle a transfer to the Sun
Fixes #126

`SSat.ks`: new script, similar to KMSat.ks. Note that the entry of very
large numbers into the required apoapsis and periapsis crashes kOS, so
distances are entered in gigameters and converted. It might be the case
that adding a `.0` on the end of the original values would have allowed
them to be read in as floats instead of integers.

`lib_transfer.ks`: `doTransfer()` was falling over if the destination
(or current body) was the Sun. Also, when transferring to the Sun, the
correction checker looks for corrections that are miniscule compared to
the size of a Solar orbit.

With the two files, I have tested this can be used to launch satellites
into a Solar orbit.

I haven't tested if Sun->Planet transfers now work, though they should
be similar enough to Planet->Moon.