Allow staging during orbit changes
The orbit change and orbit match functions have a `can_stage` parameter,
but this isn't actually useful as it still wants you to have enough
delta-v in the current stage - I have made a change so that the delta-v
check is disregarded if we are allowed to stage.

Part of a set of minor changes needed for #126

Coming soon - an additional check that we actually have more engines on
the craft.