Handle 0-eccentricity orbits in `taAt()`
Issue #126

Kerbin's orbit is a pure circle, so our `calcTa()` function, called from
`taAt()`, falls over when trying to determine where it is. In this case,
we can cheat slightly, and use `ORBIT:TRUEANOMALY` to get the current
anomaly, knowing it will advance at a constant rate over the period of
an orbit.