Delta-v calc now includes solid fuel
The delta-v calc now includes solid fuel, but only if it is a resource
within a part that is a currently-active engine i.e. it will show a more
realistic delta-v figure at launch (though not taking into account burn
time) but doesn't include non-firing engines e.g. sepratrons when
determining the delta-v for a stage.