Hold launch if 1st stage has clamps but no engines
Hold launch if 1st stage has clamps but no engines - launch is refused
until the player changes the staging layour so that the first stage has
engines.

Check yo' stagin'!

This could be expanded to check for other silliness, such as a parachute
in the first stage. It's currently possible to pass the stage
verification with no engines in the first stage, for example.

Issue #129

While I was testing this, I noticed that the launch abort sequence is a
bit keen to jettison off the escape tower while the craft is still
pitching around, so I've moved that to later in the sequence.