Don't disarm 'safe' chutes
Don't disarm 'safe' chutes that have been staged prior to re-entering.

Background: During re-entry, we check to see if parachutes were
'accidentally' staged prior to hitting the atmosphere and used to disarm
them by default. That was not really necessary for chutes that have the
deploy mode set to "when safe" as it is safe to trigger them prior to
re-entry. It's actually safer to trigger the chutes in advance this way,
in case we run out of electricity on the way down.

Now we will disarm any chutes whose deploy mode isn't "when safe" (other
possibilities are "when risky" or "immediate")

This is related to issue #112