Issue #14 further fixes
Issue #14 further fixes

* stop using LND_G_ACC as the lock appears to have been out of the
global scope, so was returning 0.
* allow the vspeed pidloop to return values below 0. Required after
making the fix above!
* landing site optimisation (slope detection) is now done outside the
doLanding() loop (previously it was only done for certain runmodes), so
it is run once each boot. This should reduce the number of times the
lander has to correct a large sideways velocity due to the site having
been moved after the deorbit burn.
* a fair bit of logging has been removed for the moment...