Various minor tweaks
Includes a new boot script for launching a crewed "satellite" - i.e.
something for which we don't have a mission boot file yet.

Added scripts rdz_dock.ks and science.ks, which can be triggered in the
same way as node.ks - to get something automated outside of a mission
file.